+++ scripts/update_public_server.sh
@@ -8,9 +8,13 @@ src_dir="/src/valhalla"
 
 git config --global --add safe.directory /src/valhalla
 
+for s in $(ls /src/valhalla/third_party); do
+    git config --global --add safe.directory /src/valhalla/third_party/$s
+done
+
 git -C "${src_dir}" checkout master
 git -C "${src_dir}" pull
-git submodule update --init --recursive
+git -C "${src_dir}" submodule update --init --recursive
 
 # remove the build folder first
 rm -r "${src_dir}"/build