+++ src/loki/search.cc
@@ -31,7 +31,6 @@ bool search_filter(const DirectedEdge* edge,
   uint32_t min_road_class = static_cast<uint32_t>(filter.min_road_class_);
   uint32_t max_road_class = static_cast<uint32_t>(filter.max_road_class_);
 
-  auto lvl = tile->edgeinfo(edge).level();
   // Note that min_ and max_road_class are integers where, by default, max_road_class
   // is 0 and min_road_class is 7. This filter rejects roads where the functional
   // road class is outside of the min to max range.
@@ -41,8 +40,7 @@ bool search_filter(const DirectedEdge* edge,
          (filter.exclude_ramp_ && (edge->use() == Use::kRamp)) ||
          (filter.exclude_closures_ && (costing.flow_mask() & kCurrentFlowMask) &&
               tile->IsClosed(edge) ||
-          (filter.level_ != kMaxLevel &&
-           std::find(lvl.begin(), lvl.end(), filter.level_) == lvl.end()));
+          (filter.level_ != kMaxLevel && !tile->edgeinfo(edge).includes_level(filter.level_)));
 }
 
 bool side_filter(const PathLocation::PathEdge& edge, const Location& location, GraphReader& reader) {