+++ src/mjolnir/osmdata.cc
@@ -7,9 +7,9 @@
 #include "filesystem.h"
 #include "midgard/logging.h"
 #include "mjolnir/osmdata.h"
-#include "mjolnir/util.h"
 
 using namespace valhalla::mjolnir;
+using valhalla::baldr::ConditionalSpeedLimit;
 
 namespace {
 
@@ -26,6 +26,7 @@ const std::string unique_names_file = "osmdata_unique_strings.bin";
 const std::string lane_connectivity_file = "osmdata_lane_connectivity.bin";
 const std::string pronunciation_file = "osmdata_pronunciation_file.bin";
 const std::string language_file = "osmdata_language_file.bin";
+const std::string conditional_speed_limit_file = "osmdata_conditional_speed_limit_file.bin";
 
 // Data structures to assist writing and reading data
 struct TempRestriction {
@@ -308,6 +309,24 @@ bool write_linguistic(const std::string& filename, const LinguisticMultiMap& lin
   return true;
 }
 
+bool write_conditional_speed_limits(const std::string& filename,
+                                    const ConditionalSpeedLimitsMultiMap& speed_map) {
+  // Open file and truncate
+  std::ofstream file(filename.c_str(), std::ios::out | std::ios::binary | std::ios::trunc);
+  if (!file.is_open()) {
+    LOG_ERROR("write_linguistic failed to open output file: " + filename);
+    return false;
+  }
+
+  uint32_t sz = speed_map.size();
+  file.write(reinterpret_cast<const char*>(&sz), sizeof(uint32_t));
+  for (const std::pair<const uint64_t, ConditionalSpeedLimit>& sp : speed_map) {
+    file.write(reinterpret_cast<const char*>(&sp), sizeof(sp));
+  }
+  file.close();
+  return true;
+}
+
 bool read_restrictions(const std::string& filename, RestrictionsMultiMap& res_map) {
   // Open file and truncate
   std::ifstream file(filename, std::ios::in | std::ios::binary);
@@ -525,6 +544,28 @@ bool read_linguistic(const std::string& filename, LinguisticMultiMap& ling_map) 
   return true;
 }
 
+bool read_conditional_speed_limits(const std::string& filename,
+                                   ConditionalSpeedLimitsMultiMap& speed_map) {
+  // Open file and truncate
+  std::ifstream file(filename, std::ios::in | std::ios::binary);
+  if (!file.is_open()) {
+    LOG_ERROR("read_conditional_speed_limits failed to open input file: " + filename);
+    return false;
+  }
+
+  // Read the count and following conditional speed limits
+  uint32_t count = 0;
+  file.read(reinterpret_cast<char*>(&count), sizeof(uint32_t));
+  speed_map.reserve(count);
+  while (count--) {
+    std::pair<uint64_t, ConditionalSpeedLimit> sp = {{}, {}};
+    file.read(reinterpret_cast<char*>(&sp), sizeof(sp));
+    speed_map.emplace(std::move(sp));
+  }
+  file.close();
+  return true;
+}
+
 } // namespace
 
 namespace valhalla {
@@ -554,17 +595,19 @@ bool OSMData::write_to_temp_files(const std::string& tile_dir) {
   file.close();
 
   // Write the rest of OSMData
-  bool status = write_restrictions(tile_dir + restrictions_file, restrictions) &&
-                write_viaset(tile_dir + viaset_file, via_set) &&
-                write_access_restrictions(tile_dir + access_restrictions_file, access_restrictions) &&
-                write_bike_relations(tile_dir + bike_relations_file, bike_relations) &&
-                write_way_refs(tile_dir + way_ref_file, way_ref) &&
-                write_way_refs(tile_dir + way_ref_rev_file, way_ref_rev) &&
-                write_node_names(tile_dir + node_names_file, node_names) &&
-                write_unique_names(tile_dir + unique_names_file, name_offset_map) &&
-                write_lane_connectivity(tile_dir + lane_connectivity_file, lane_connectivity_map) &&
-                write_linguistic(tile_dir + pronunciation_file, pronunciations) &&
-                write_linguistic(tile_dir + language_file, langs);
+  bool status =
+      write_restrictions(tile_dir + restrictions_file, restrictions) &&
+      write_viaset(tile_dir + viaset_file, via_set) &&
+      write_access_restrictions(tile_dir + access_restrictions_file, access_restrictions) &&
+      write_bike_relations(tile_dir + bike_relations_file, bike_relations) &&
+      write_way_refs(tile_dir + way_ref_file, way_ref) &&
+      write_way_refs(tile_dir + way_ref_rev_file, way_ref_rev) &&
+      write_node_names(tile_dir + node_names_file, node_names) &&
+      write_unique_names(tile_dir + unique_names_file, name_offset_map) &&
+      write_lane_connectivity(tile_dir + lane_connectivity_file, lane_connectivity_map) &&
+      write_linguistic(tile_dir + pronunciation_file, pronunciations) &&
+      write_linguistic(tile_dir + language_file, langs) &&
+      write_conditional_speed_limits(tile_dir + conditional_speed_limit_file, conditional_speeds);
   LOG_INFO("Done");
   return status;
 }
@@ -609,7 +652,8 @@ bool OSMData::read_from_temp_files(const std::string& tile_dir) {
       read_unique_names(tile_directory + unique_names_file, name_offset_map) &&
       read_lane_connectivity(tile_directory + lane_connectivity_file, lane_connectivity_map) &&
       read_linguistic(tile_directory + pronunciation_file, pronunciations) &&
-      read_linguistic(tile_directory + language_file, langs);
+      read_linguistic(tile_directory + language_file, langs) &&
+      read_conditional_speed_limits(tile_directory + lane_connectivity_file, conditional_speeds);
   LOG_INFO("Done");
   initialized = status;
   return status;