+++ src/worker.cc
@@ -410,7 +410,7 @@ void parse_location(valhalla::Location* location,
     location->mutable_search_filter()->set_exclude_ramp(
         rapidjson::get<bool>(*search_filter, "/exclude_ramp", false));
     location->mutable_search_filter()->set_level(
-        rapidjson::get<int>(*search_filter, "/level", static_cast<int>(baldr::kMaxLevel)));
+        rapidjson::get<float>(*search_filter, "/level", static_cast<float>(baldr::kMaxLevel)));
     // search_filter.exclude_closures
     exclude_closures = rapidjson::get_optional<bool>(*search_filter, "/exclude_closures");
   } // or is it pbf