+++ valhalla/baldr/graphconstants.h
@@ -209,8 +209,11 @@ constexpr uint32_t kMaxAddedTime = 255;
 // this is the minimum we support, i.e. -500 m would result in "no elevation"
 constexpr float kNoElevationData = -500.0f;
 
-// Maximum level value
-constexpr int16_t kMaxLevel = std::numeric_limits<int16_t>::max();
+// (building) level constants
+// highest 3-byte value
+constexpr float kLevelRangeSeparator = 1048575.0f;
+constexpr float kMinLevel = std::numeric_limits<float>::min();
+constexpr float kMaxLevel = std::numeric_limits<float>::max();
 
 // Node types.
 enum class NodeType : uint8_t {
@@ -385,9 +388,11 @@ enum class TaggedValue : uint8_t { // must start at 1 due to nulls
   kLayer = 1,
   kLinguistic = 2,
   kBssInfo = 3,
-  kLevel = 4,
+  kLevel = 4, // deprecated in favor of kLevels
   kLevelRef = 5,
   kLandmark = 6,
+  kConditionalSpeedLimits = 7,
+  kLevels = 8,
   // we used to have bug when we encoded 1 and 2 as their ASCII codes, but not actual 1 and 2 values
   // see https://github.com/valhalla/valhalla/issues/3262
   kTunnel = static_cast<uint8_t>('1'),